Move transaction attributes to Abstractions
